+++ tests/integration/aliases/aliases_test.go
@@ -4,7 +4,6 @@ package ints
 
 import (
 	"path"
-	"path/filepath"
 	"testing"
 
 	"github.com/pulumi/pulumi/pkg/testing/integration"
@@ -34,29 +33,5 @@ func TestNodejsAliases(t *testing.T) {
 				},
 			})
 		})
-	}
-}
-
-func TestPythonAliases(t *testing.T) {
-	for _, dir := range dirs {
-		if dir == "adopt_into_component" {
-			d := path.Join("python", dir)
-			t.Run(d, func(t *testing.T) {
-				integration.ProgramTest(t, &integration.ProgramTestOptions{
-					Dir: path.Join(d, "step1"),
-					Dependencies: []string{
-						filepath.Join("..", "..", "..", "sdk", "python", "env", "src"),
-					},
-					Quick: true,
-					EditDirs: []integration.EditDir{
-						{
-							Dir:             path.Join(d, "step2"),
-							Additive:        true,
-							ExpectNoChanges: true,
-						},
-					},
-				})
-			})
-		}
 	}
 }
